# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WealthApp API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pathway/{stepNumber}:
    put:
      operationId: upsertPathwayStep
      tags: [pathway]
      summary: Save pathway step progress
      parameters:
        - name: stepNumber
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertPathwayStepBody"
      responses:
        "200":
          description: Updated step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PathwayStep"
...
  /assets:
    get:
      operationId: listAssets
      tags: [networth]
      summary: List assets for current user
      responses:
        "200":
          description: Assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [networth]
      summary: Create an asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    put:
      operationId: updateAsset
      tags: [networth]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    delete:
      operationId: deleteAsset
      tags: [networth]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /liabilities:
    get:
      operationId: listLiabilities
      tags: [networth]
      summary: List liabilities for current user
      responses:
        "200":
          description: Liabilities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Liability"
    post:
      operationId: createLiability
      tags: [networth]
      summary: Create a liability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLiabilityBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Liability"
  /liabilities/{id}:
    put:
      operationId: updateLiability
      tags: [networth]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLiabilityBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Liability"
    delete:
      operationId: deleteLiability
      tags: [networth]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /plans/{id}:
    get:
      operationId: getFinancialPlan
      tags: [plans]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialPlan"
    put:
      operationId: updateFinancialPlan
      tags: [plans]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFinancialPlanBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialPlan"
  /plans/{id}/milestones:
    get:
      operationId: listPlanMilestones
      tags: [plans]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlanMilestone"
    post:
      operationId: createPlanMilestone
      tags: [plans]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanMilestoneBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanMilestone"
...
  /conversations:
    get:
      operationId: listConversations
      tags: [messages]
      summary: List conversations
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [messages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /conversations/{id}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
  /advisor/clients/{id}/plan:
    get:
      operationId: advisorGetClientPlan
      tags: [plans]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialPlan"
        "404":
          description: No plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: advisorCreateClientPlan
      tags: [plans]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFinancialPlanBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialPlan"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    PathwayStep:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        stepNumber:
          type: integer
        stepName:
          type: string
        status:
          type: string
        formData:
          type: object
          nullable: true
        completedAt:
          type: ["string", "null"]
        updatedAt:
          type: string
      required: [id, userId, stepNumber, stepName, status, updatedAt]
    UpsertPathwayStepBody:
      type: object
      properties:
        stepName:
          type: string
        status:
          type: string
        formData:
          type: object
      required: [stepName, status]
...
    Asset:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        category:
          type: string
        valueUsd:
          type: string
        currencyOriginal:
          type: string
        valueOriginal:
          type: ["string", "null"]
        asOfDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, name, category, valueUsd, currencyOriginal, createdAt, updatedAt]
    CreateAssetBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        valueUsd:
          type: string
        currencyOriginal:
          type: string
        valueOriginal:
          type: string
        notes:
          type: string
      required: [name, category, valueUsd]
    Liability:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        category:
          type: string
        balanceUsd:
          type: string
        interestRatePercent:
          type: ["string", "null"]
        monthlyPaymentUsd:
          type: ["string", "null"]
        currencyOriginal:
          type: string
        balanceOriginal:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, name, category, balanceUsd, currencyOriginal, createdAt, updatedAt]
    CreateLiabilityBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        balanceUsd:
          type: string
        interestRatePercent:
          type: string
        monthlyPaymentUsd:
          type: string
        currencyOriginal:
          type: string
        balanceOriginal:
          type: string
      required: [name, category, balanceUsd]
...
    FinancialPlan:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        advisorId:
          type: ["string", "null"]
        title:
          type: string
        status:
          type: string
        planData:
          type: ["string", "null"]
        nextReviewDate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, clientId, title, status, createdAt, updatedAt]
    UpdateFinancialPlanBody:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
        planData:
          type: string
        nextReviewDate:
          type: string
      required: [title]
    PlanMilestone:
      type: object
      properties:
        id:
          type: string
        planId:
          type: string
        title:
          type: string
        targetDate:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, planId, title, status, createdAt]
    CreatePlanMilestoneBody:
      type: object
      properties:
        title:
          type: string
        targetDate:
          type: string
        status:
          type: string
        notes:
          type: string
      required: [title]
...
    Conversation:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: ["string", "null"]
        advisorId:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, createdAt]
    CreateConversationBody:
      type: object
      properties:
        clientId:
          type: string
        advisorId:
          type: string
    Message:
      type: object
      properties:
        id:
          type: string
        conversationId:
          type: string
        senderId:
          type: ["string", "null"]
        senderRole:
          type: ["string", "null"]
        content:
          type: string
        attachmentUrl:
          type: ["string", "null"]
        sentAt:
          type: string
        readAt:
          type: ["string", "null"]
      required: [id, conversationId, content, sentAt]
    SendMessageBody:
      type: object
      properties:
        content:
          type: string
        senderRole:
          type: string
        attachmentUrl:
          type: string
      required: [content]
...